{{ cookiecutter.project_slug }}/requirements/local.txt
-r base.txt
-e git+https://github.com/jazzband/django-debug-toolbar.git#egg=django-debug-toolbar
factory_boy==2.12.0
pytest==5.3.2
pytest-django==3.7.0
